feat(student-dashboard): render tags and dates
